requirements.txt
numba
librosa
numpy==1.23.3
scipy
torch
unidecode
openjtalk>=0.3.0.dev2
jamo
pypinyin
jieba
protobuf
cn2an
inflect
eng_to_ipa
ko_pron
indic_transliteration
num_thai
opencc
audonnx
openai
tencentcloud-sdk-python
pyaudio
simpleaudio
pydub
webrtcvad
asyncio
aiohttp
pvporcupine
paho-mqtt
langchain
pinecone-client
google-api-python-client
keyboard
azure-cognitiveservices-speech
requests
clueai